Allow task parameters to be set when using the run command.
Currently there is a 'task_params' option which can add to or override
values in the params dictionary for a task, but it is only available
when running a task with 'airflow test'.

By accepting this parameter in 'airflow run' and then passing it to the
subprocess through the command method in the TaskInstance class this
option can be supported.

This has use cases in running tasks in an ad-hoc manner where a
parameter may define an environment (i.e. testing vs. production) or
input / output locations and a developer may want to tweak them on the
fly.